fix footer styling and rename logout button to profile button
