+++ src/components/footer/Footer.module.scss
@@ -1,11 +1,9 @@
 .footer {
-  position: fixed;
-  bottom: 0;
-  left: 0;
+  box-sizing: border-box;
   display: flex;
   justify-content: space-between;
-  width: 100vw;
-  padding: $main-content-padding;
+  padding: $main-content-padding $container-padding;
+  margin: #{$container-padding} #{-$container-padding} 0 #{-$container-padding};
   color: white;
   background-color: #00000066;
