--- src/components/navigation/logout-button/index.tsx
@@ -1,3 +1,0 @@
-import LogoutButton from './LogoutButton';
-
-export default LogoutButton;
