+++ src/components/navigation/profile-button/index.tsx
@@ -1,0 +1,3 @@
+import LogoutButton from './ProfileButton';
+
+export default LogoutButton;
