+++ src/styles/globals.scss
@@ -16,8 +16,17 @@ a {
   text-decoration: none;
 }
 
-.container {
-  padding: 0 2rem;
+// Do not re-use elsewhere
+.app-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  min-height: 100vh;
+  padding: 0 $container-padding;
+}
+
+.page-content {
+  flex-grow: 1;
 }
 
 * {
