+++ src/styles/variables.scss
@@ -64,3 +64,5 @@ $main-content-br: 10pt;
 
 $card-item-br: 5pt;
 $card-item-padding: 6pt;
+
+$container-padding: 2rem;
